qdev: change qbus_walk_child_* to qbus_walk_children
Sometimes we want to iterate both devices in busses in the same pass so fold
the two walk functions into one.

Paolo suggested that we model the return values on how GCC's walkers work which
allows an actor to skip child transversal, or terminate walking with a positive
value that's returned as the qbus_walk_children's result.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>