vmstate: change VMStateField to take a string instead of VMStateInfo
VMStateInfo is the mechanism that's used to save a field.  VMStateField and
VMStateDescription, however, describe the data format and should not directly
introduce a dependence on how individual fields are saved.

Besides being a conceptual clean-up, this reduces the build dependency by allow
vmstate to exist without any build dependency on savevm.c

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>