qdev: eliminate ->id and replace it with ->name, ->name is mandatory
This breaks everything so more work is needed.

Device's have to have a unique name and id is a giant hack.  The reason device's
need to have a name is that we need a consistent way to identify a device on a
bus spanning two instances of QEMU.  The order of devices on a bus is
meaningless today so in order to migrate from one instance of QEMU to another,
we need to add a meaningful order when we migrate.

By having a universal name, we can sort by the name giving us a stable order.
The catch is that name has to be stable across QEMU invocations.

But QEMU can automate an awful lot of this.  For devices that only have one
instance, the board creation can specify a default name.  For busses that have
some sort of unique identifier already, the bus may default the name of the
device.

But ultimately, this is going to be user facing because there are going to be
certain cases where a user is forced to specify the name.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>