cow2: introduce a L2 cache
The L2 cache should improve read performance.  The cache is a write-through
cache.  In order to support the cache, we need to explicitly allocate entries
from the cache, ensure they are coherent with what's on disk, and then
explicitly commit the entry to the cache.

When searching for cached entries, we only look at committed entries.  When
using cached entries, a caller must explicitly free an entry when it no longer
needs it.  Internally, we use reference counting to deal with pruning
non-committed entries.

For now, we set the maximum cache size to 1.  This is useful for debugging since
it should behave exactly as the previous code.

The next commit will bump the cache size which should help bisecting in the
case that there's a bug in the life cycle management.

We still keep the notion of current L2 entry but there's no reason why the L2
cache would prohibit multiple simultaneous requests.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>