cow2: get rid of cur_l2_table
Which eliminates the dependence on global state for each request.  We do this
by introducing a new Cow2Request structure that is associated with an AIOCB.
This contains the global state a request needs.

We have to put it in a separate structure because certain functions don't
actually take an AIOCB because they can be called outside of the aio calls.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>