qed: make some changes to the image format
Rename the magic from cow2 -> qed\0

Introduce compat features.  These are features that don't affect the ability
to read an image file so if an unsupported compat feature is encountered, it
can be safely ignore.

Introduce a first_cluster.  This is a cluster index of the first usable cluster.
When parsing the image and performing offline manipulation, because of compat
features it's not possible to know the header size.  first_cluster provides
information about the header size indirectly.

Make backing_file a feature.  This is mostly useful as a demonstration of how
to introduce format features.

Making backing_fmt a compat feature.  Right now, we enforce that a backing_fmt
is specified.  This really isn't neccessary since format is really information
to qemu.  One can imagine a different tool that only supported cow2 and did
not suffer from the problem with raw images that qemu does.

This is done in one commit to try to avoid lots of commits with format changes.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>