qed: use async_context_{push,pop}
Async contexts prevent bottom halves that were registered in a previous context
from being invoked while running qemu_aio_wait().  When we implement blocking
functions in qed, we don't want any bottom halves to be executed while we block
because it could lead to re-entrance in functions that don't support it.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>