qed: detect zero writes and skip them when to an unalloc cluster
A value of 1 is used to indicate that a cluster contains all zeros.  Update the
code to detect zero writes only when a flag is set on the AIOCB.  For now, only
set the flag on copy-on-read based write requests to avoid polluting the
cache on write in the zero copy case.

After this patch, we can stream an image file from a backing file without
fully expanding the image.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>