qed: fix race when committing an L2 cache entry
When trying to add an entry to the L2 cache, a caller has to allocate an
uncommitted cache entry, read the contents of the entry, and then commit the
entry to the cache.

However, if two requests are both trying to satisfy the same L2 cache entry,
one read() may take arbitrarily long whereas the other entry may be read,
updated, and rewritten before the previous read() completes.  When the second
entry is committed, after the very long read, we simply avoid committing the
entry.

The code today assumes that after committing an L2 cache entry, that entry is
present in the cache but as illustrated above, this is not necessarily the case.
Furthermore, the entry not present in the cache may actually be out of date
leading to data loss.

The solution is to avoid assuming the L2 cache entry is valid after commit and
steal a reference to the L2 cache entry in the commit() function to force a
caller to get a new entry to it.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>