qed: Autoclear feature bits
Avi Kivity <avi@redhat.com> suggested a backwards compatibility
mechanism that allows image files to degrade gracefully when opened by
an older program.  Autoclear feature bits behave like compat feature
bits except that unknown bits are reset when opening the image file.

This allows a newer program to detect that the feature bit has been
reset and is no longer consistent.  The newer program should then clean
up the feature and may reinstate the feature again.

The rule of thumb is:
1. Feature bits should be used for incompatible changes to the file
   format.
2. Compat feature bits should be used for extensions to the file format
   which can be safely ignored by older programs and do not require
   maintaining consistency with the image data.
3. Autoclear feature bits should be used for extensions to the file
   format which can be safely ignored but once ignored are no longer
   consistent with the image data.

An example of an autoclear feature is a cluster freelist which can be
rebuilt during consistency check if the autoclear bit has been reset.

Signed-off-by: Stefan Hajnoczi <stefanha@linux.vnet.ibm.com>