virtproxy: modify virtproxy base to accept a chardev
This allows us to pass a chardev to virtproxy instead of a listening FD.
This also breaks out the virtproxy packet processing routines so they can
be used directly when the device calls the chardev write handler,
and wraps writes to the guest with a function that calls the device's
write handler, or just writes to the channel fd, depending on the
context (chardev (host, qemu) vs. channel fd (guest, qemu-vp)).