virtproxy: integration, qemu-vp fix-ups
Modify the qemu-vp guest daemon to use new virtproxy constructor. Also
noop vp_chr_read (our qemu_chr_read wrapper for when running on the host
as a chardev) since we don't use a chardev in the guest and since
handling this with #ifdefs would doing a reconfigure on the build tree.