virtagent: don't spin on socket when read()=0
Ideally we'd close() at this point, but in our case it's expected that
the socket always remain open (host-side chardev), or is being used to
pipe in data from a serial port. Will revisit how to handle this when
isa-serial support no longer depends on something like socat as an
in-between.