Fix pid file handling by using file locks (V2)
Changes since V1:
- Switched to using straight, unbufferred fds
- Switched to lockf to remove <sys/file.h> dependency

The current qemu-va code does not provide a facility for cleaning up stale pid
files.  As such, the program will fail to start on subsequent invocations
unless the pid file is first removed.  This is obviously incorrect behavior and
can be easily fixed by using advisory file locks.

We open the pid file (creating it if it does not exist) and try to lock
it.  If this is successful, we can write our pid and continue executing
(while keeping the pid file open).  When the daemon exits, the file lock
will expire.

Signed-off-by: Adam Litke <agl@us.ibm.com>
--