qapi: fix reporting of balloon size
The normal behavior of query-balloon is to report the size of usable guest
memory (in bytes), not the current size of the balloon.  The human monitor
reports the size of guest memory in MB.  This patch fixes a regression
introduced in 05ab3085c687e69ecb1b66028cf940dc6fc89d13 that was causing the
size of the balloon to be reported instead.

Signed-off-by: Adam Litke <agl@us.ibm.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>