qapi: fix race in vnc event test
I had thought a previous commit introduced a race, but as it turns out, the test
case was racey.  We can't rely on QEMU delivering events in a timely fashion
so assumptions about events being delivered after a wait_event call are just
flat wrong.  We can get away with it in the cont/stop tests because those events
are sent on the same socket as QMP but since the VNC tests are a different
connection, there's just no idea as to when the QMP event will come in relation
to data on the VNC session.