qapi: refactor getfd/closefd to be stateful commands
We still don't really have a good story with regards to supporting these
commands.  I think we can probably support the QMP client interface pretty
easily by having something like a libqmp_session_add_fd(sess, fd) which adds
a fd to be sent with the next command.

However, that's not usable internally so there's simply no way to make
the HMP to QMP bridge work in this model.  Requires some careful thought and
may not be possible to make 100% transparent.