vga-isa: fix migration by breaking it
This is pretty sad.  We use the same section name for vga-isa as we do for
vga-pci even though we use separate formats.  This breaks the live migration
protocol because we may misinterpret the vga-isa as a vga-pci device.

vga-isa should use it's own wrapper just like vga-pci does.  That's what we do
in this patch.

Signed-by-off: Anthony Liguori <aliguori@us.ibm.com>