fdc: fix migration of non-ISA fdc devices
Again, we are not using a unique name for non-ISA fdc devices and as a result,
we may misinterpret an ISA fdc device as a non-ISA fdc device.

In this case, both ISA fdc and non-ISA fdc are using too generic of a name.
Since ISA fdc is more common, let it keep the old name.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>