json-parser: increase parser speed by 4100%
The recursive decent parser we use needs to keep track of the current position
in the token stream.  Since there was no way to store an iterator for a QList,
we just copied the list in it's entirety and passed around new versions of the
list.  For very large token streams, the overhead of this was huge.

By introducing a qlist iterator and using that instead of copying lists, we
increase the speed of the test-vmstate (which uses a large JSON file) from
2.83 seconds to 0.07 seconds.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>