piix4: only allow guest instantiated device unplug on explicitly marked devices
When the unplug path is broken, a guest can potentially trigger bad behavior
by initating a device unplug.  Prevent this by allow devices to opt-in to
guest initiated unplug which should err on the side of being conservative.

Right now, we just add virtio devices to the white list because they are the
only devices that I know are consistently tested with unplug.  We can add more
as we see fit.

Reported-by: Nelson Elhage <nelhage@ksplice.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>