This is a simple tool that I've been using for the past couple months to help with day-to-day testing of changes.  It may seem like it's similar to kvm-autotest but it's actually quite different.  Most noticably:
 - It's a coverage test instead of an acceptance test.  Each time it runs it
   uses a slightly a semi-random configuration for the guest.  This means that
   the more you run it, the more coverage you get.  This is a good fit for
   maintainer testing because you get wide testing without having to wait for
   48-hour test cycles.

 - It runs in the foreground as an unprivileged user.  This means I can kick it
   off in a terminal while I go off and do something else, but still keep an
   eye on what's going on.

 - It works with TCG guests too and includes a TCG test case for the pseries
   Power machine.

That said, it's *not* a replacement for KVM autotest.  It is not a good tool
for doing acceptance testing, like you would do to cut a new QEMU release.  But
perhaps there's some behavior that could be leveraged by KVM autotest in the
future here.

I'm not proposing this for tree inclusion right now.  Just sharing a tool that
I've found to be useful.  I really just want the previous patch to go in so that
I can stop carrying that patch privately.

Right now, you need to setup an ISO directory to use the test tool.  After
copy-on-read lands in the tree, I plan on making it create CoR files backing to
http so that no explicit setup is required.

There's no point in running each distro test since they're all basically testing
the same thing.  Instead, randomly choose which Linux distro to test.  This
better suites the nature of the test suite as something that is pretty light
weight to run but scales its coverage as you run it over and over again.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>