qom: add Plug class
The Plug class is the base class for all objects.  Plugs implement properties
using the existing Visitor framework.  This allows for very complex properties
to be implemented.

Plugs also implement life cycle management by containing a realized state.  An
object isn't fully constructed until its realized.  By setting realized to
false, an object can also be reset.

Plug properties can be set to be read-only or read-write.  A read-write property
can be locked.  While a read-write property is locked, it cannot be written to.

The expectation is that some properties will be locked at realize time.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>