plug: add socket property type
Sockets form back links in the object graph.  From the object's perspective,
a socket just looks like a pointer to an object.  The socket property type
provides a generic mechanism to set those pointers to other objects while
providing strong type checking.

It's also useful to lock sockets, particularly after realize.  This allows for
an object to enforce that a socket is programmed prior to realize and then not
modified afterwards.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>