plug: add Plug property type
Plug properties types allow composition within the object model.  A plug is an
object that is directly created from the current object, usually using the same
memory as the current object.

Realize state is propagated to any plugs assocatied with the object.  Lifecycle
is also propagated such that when a plug is added, its automatically initialized
and the plug is finalized when the parent object is finalized.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>