char: introduce backend tx queue
While the front tx queue has no flow control, the backend tx queue uses a
polling function to determine when the front end can receive data.

To convert this to the new queue model, we simply try to flush the backend tx
queue whenever we poll.  We then return the remaining space in the queue as
the value of the polling function.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>