Scripts to generate project stats
As part of my talk for KVM Forum, I am collecting some stats on the project
since last year.  I thought I'd share the scripts in case anyone is interested
in how they work.

I think this is just about all of the data I need, but patches are certainly
welcome.

Of course, you'll have to come to KVM Forum to see the pretty version of these
stats (there should be videos too for those that can't make it :-))

And thanks to Alex for poking me to collect these too.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>