add support for machine models to specify their migration format
We need to provide a new migration format, and not break migration
in old machine models.  So add a migration_format field to QEMUMachine.

This adds a dependency on ram_addr_t to savevm.c (through hw/boards.h),
whose type definition depends on the word size.  So move savevm.o to
libhw.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>