[RFC] time: refactor QEMU timer to use GHRTimer
This replaces all of the QEMU timer code with GHRTimer, dramatically simplifying
time keeping in QEMU while making it possible to use QEMUTimer code outside of
the main loop.  The later is critical to building unit tests.

This is an RFC because I'm sure this breaks things as it changes things.  QEMU
time keeping is quite a mess today.  Here's what we do today:

1) We have three clocks:
  a) the real time clock, based on system time, not monotonic
  b) the host clock, based on the real time clock, monotonic by detecting
     movements backward in time
  c) the vm clock, based on real time clock but may start/stop with the guest

2) A "cpu ticks" clock that uses platform specific mechanisms (inline asm)

3) Various clock source implementations that may use a periodic timer or a
   a dynamic time source.  We have different implementations for different
   platforms

4) Time events are delivered via SIGALRM which means we end up getting EINTRs
   very often in QEMU.  This is fairly annoying.  Signals also race against
   select leading to a very ugly set of work arounds involving writing data to
   pipes.  This is the sort of stuff in Unix programming that I wish I never had
   to learn about and am very eager to eliminate in QEMU :-)

(2) is just plain broken.  In modern operating systems, gettimeofday() is
optimized CPU instructions when they can be used safely.  Often they can't be
used safely and we ignore that in QEMU.  For instance, on x86, RDTSC races with
the scheduler (not to mention that the TSC is infamously unstable across cores).
The kernel does the right thing here and provides the fastest method that's
correct.

(1.a) seems like a bug more than a feature.  I don't see a lot of disadvantages
to using a monotonic time source.

(1.b) is a bit naive in its current form.  Modern kernels export a truly
monotonic time source which has a reliable frequency.  Even though (1.b) detects
backwards jumps, it doesn't do anything about large forward jumps which can also
be problematic.

(3) results in tons of bitrot.  The only implementation that really should be
used these days is dynticks.

(4) is just painful

I've tested TCG, KVM, and icount.  They all seem to be working as well as they
did before.  I've only tested a Linux host though.  I'm sure I may be breaking
something here and am hoping folks can help identify it.

This depends on the I/O thread and glib patches I posted earlier today.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>