qdev: take ownership of id pointer
qdev has this quirk that it owns a seemingly arbitrary QemuOpts pointer.  That's
because qdev expects a static string for the id (which really makes no sense
since ids are supposed to be provided by the user).  Instead of managing just
the id pointer, we currently take ownership of the entire QemuOpts structure
that was used to create the device just to keep the name around.

Just strdup the pointer we actually need.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>