qdev: add ability to do QOM-style derived naming
By using a prefix of "::" in the name, we can safely derive the composed device
name from the parent device and busses name.  For instance, if the "::i440fx"
device created a device named "piix3", it would look like this:

 static void i440fx_initfn(...)
 {
    s->piix3 = qdev_create("PIIX3", "::piix3");
    ...

The resulting device would be named "::i440fx::i440fx.0::piix3".  The reason for
the middle "::i440fx.0" blob is that there are two levels of the tree hierarchy
here and the bus level already has it's name derived from the parent device.

We'll eliminate the bus level of the hierarchy in due time, but for now we have
to just live with the ugly names.

This patch lets qdev names be specified as a printf style format string which is
convenient for creating devices like "::smbus-eeprom[%d]".

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>