spapr: provide default names to qdev devices
spapr already had logic to do default device naming but unfortunately this logic
can (and will) conflict with user supplied names.

This patch uses the same convention but uses the composition namespace to do it
safely. Instead of doing module@reg, this does module[reg] which is more in line
with how other devices currently work.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>