migration: add a MAINTAINERS entry for migration
I think this is an accurate reflection of the state of migration today.  This
is the second release in a row where we're scrambling to fix a critical issue
in migration.

The first step in fixing this problem is being up front about what the current
state of the subsystem is.  If we're going to treat migration as a release
blocking feature in the future, than we need to promote the migration subsystem
above 'Odd Fixes' status.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>