pyembed: integer python into QEMU
This is something I started during 1.0-rc on a lark and spent some time last
night actually making work.  I'm sending it only to show that (1) it's possible
and (2) to get some input about what other people think as a longer term
direction.

I see a couple possibilities here:

 1. We could code portions of QEMU as call-outs to Python.  Command line
    parsing, GUI interface, etc. are all good candidates to be moved to Python.

 2. We could support loading of user-supplied Python modules.  This would be
    convenient as an extension mechanism since a HIL binding provide a clean
    interface that we could make more stable than just loading a random .so or
    including a .c file into the build.

 3. We could use this as an super tracing mechanism.  I can imagine using
    something like this in a production environment to trouble shoot a complex
    problem.

I'm not totally decided here myself on whether any of this is a good idea but
since this has come up so many times in the past, I figured it's at least worth
exploring.