safe-buffer: add bound checked buffer class
In an effort to be more defensive, add a bounds checked buffer class.  This
class provides a safe way to allocate, grow, copy into, out of, and get a
direct pointer to a buffer.  It's safe against underflows, overflows, and
integer overflow attacks.

I tried to make it as easy to use as possible and so far, in practice, it ends
up simplifying code.  There's some trickery included that allows a completely
static SafeBuffer to be declared which means that SafeBuffers can replace
static buffers without the need to add a cleanup routine.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>