e1000: convert to use SafeBuffers
This is a straight forward conversion that, at least for me, improves
readability quite a bit.

The only interesting bit is that there was a purposeful overflow used as an
optimization (where the address of vlan was used for the full packet).  This was
subtle at best and dangerous at worst.  In this case, simply copy to a temporary
buffer instead of overflowing vlan.

Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>